some of Maphandler warnings fixed @akinsoen @celkenze @jbarber07 @mmassime

+++ testconnexion/game_test/MapHandler.hpp
@@ -86,7 +86,7 @@ class Projectile final: public MapObject{
 public:
     Projectile()=default;
     void move() override;
-    Projectile(uint_fast16_t nx, uint_fast16_t ny,int dam,bool ship_t, int h, int p) {pos.x = nx; pos.y = ny; damage=dam; shipType=ship_t;typ=projectile; hp = h; player = p;}
+    Projectile(int nx, int ny,int dam,bool ship_t, int h, int p) {pos.x = static_cast<uint_fast16_t>(nx); pos.y = static_cast<uint_fast16_t>(ny); damage=dam; shipType=ship_t;typ=projectile; hp = h; player = p;}
     int getDamage() const{return damage;}
     bool getShipType() const{return shipType;}
     int getPlayer() const{return player;}
@@ -97,7 +97,7 @@ class Bonus final: public MapObject{
     bonusType bonustype{};
 public:
     Bonus(){hp=20;};
-    Bonus(uint_fast16_t nx, uint_fast16_t ny,bonusType bonus_t) :bonustype(bonus_t)  {pos.x = nx; pos.y = ny; hp=10;};
+    Bonus(int nx, int ny,bonusType bonus_t) :bonustype(bonus_t)  {pos.x = static_cast<uint_fast16_t>(nx); pos.y = static_cast<uint_fast16_t>(ny); hp=10;};
     bonusType getBonusType() const {return bonustype;}
     ~Bonus(){};
